requirements.txt
fastapi
uvicorn[standard]
pandas
requests
python-dotenv
comtradeapicall  # bu PyPI'da varsa, ismi bu şekilde olmalı